fix threads as int

+++ config/autoload/soluble-mediatools.global.php
@@ -3,7 +3,7 @@
 return [
     'soluble-mediatools' => [
         // do not set any threads: 0 means all cores
-        'ffmpeg.threads'   => getenv('FFMPEG_THREADS'),
+        'ffmpeg.threads'   => (int) getenv('FFMPEG_THREADS'),
         'ffmpeg.binary'  => getenv('FFMPEG_BIN'),
         'ffprobe.binary' => getenv('FFPROBE_BIN'),
     ],
